requirements.txt.txt
streamlit
tensorflow
opencv-python-headless
Pillow
numpy